api/requirements.txt
mypy
pytest
faker
pydantic[email]
fastapi
python-multipart
uvicorn[standard]
gunicorn
requests
sqlalchemy
sqlalchemy-utils
pybigquery
pymysql
psycopg2-binary
pyyaml
google-cloud-secret-manager
itsdangerous
authlib
httpx
fastapi_jwt_auth
